refactor(backend): decouple server logic into specialized routers and services
- Extract magic constant TWITCH_SESSION_ID to src/core/constants.py
- Move Twitch queue and LLM background logic to TwitchService
- Extract model endpoints out of run_server.py into model_router.py
- Fix test suite hanging issues caused by uncancelled background tasks in FastAPI lifespan
- Update related unit tests for Twitch logic and interaction priority
