+++ src/core/constants.py
@@ -1,0 +1,10 @@
+"""
+System-wide constants for Parcera Backend to prevent magic words and duplicated strings.
+"""
+
+# Session Identification
+TWITCH_SESSION_ID = "twitch-session"
+
+# Defaults
+DEFAULT_UVICORN_PORT = 8676
+DEFAULT_UVICORN_HOST = "127.0.0.1"
