+++ src/run_server.py
@@ -1,35 +1,27 @@
 import asyncio
-import json
 import logging
 import os
-import threading
-from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
 from dotenv import load_dotenv
 import httpx
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import StreamingResponse
 from aiavatar.adapter.websocket.server import AIAvatarWebSocketServer
 from aiavatar.sts.session_state_manager import SQLiteSessionStateManager
 from aiavatar.sts.performance_recorder.sqlite import SQLitePerformanceRecorder
 from core.avatar import ParceraAvatarBase
-from core.download import LoggingTqdm, check_model_cached, download_model_with_progress
 from core.engine import TTSEngineManager
 from core.config import load_config_file
 from routers.config_router import create_config_router
 from routers.tts_router import create_tts_router
 from routers.twitch_router import create_twitch_router
+from routers.model_router import create_model_router
+from services.twitch_service import TwitchService
 
-import re
-from aiavatar.sts.models import STSRequest
-from core.chat_logger import chat_logger
 from core.interaction import InteractionController
+from core.constants import TWITCH_SESSION_ID, DEFAULT_UVICORN_PORT, DEFAULT_UVICORN_HOST
 
 logger = logging.getLogger(__name__)
-
-# Constants
-TWITCH_SESSION_ID = "twitch-session"
 
 class InternalWebSocket:
     """A virtual WebSocket bridge that satisfies aiavatar's requirements and broadcasts responses to the UI."""
@@ -58,6 +50,7 @@ class ParceraServer(ParceraAvatarBase):
 
         # Interaction Controller (Orchestrates the pipeline)
         self.controller = InteractionController(self, self.config)
+        self.twitch_service = TwitchService(self)
 
         # Track current providers for hot-swapping
         self.current_stt_provider = self.config.get("stt", {}).get("provider", "faster_whisper")
@@ -83,8 +76,6 @@ class ParceraServer(ParceraAvatarBase):
             debug=self.config.verbose,
             voice_recorder_enabled=False
         )
-
-        self.twitch_queue = asyncio.Queue()
 
         # Attach callbacks to controller
         if hasattr(self.stt, "on_recognized_callback"):
@@ -147,130 +138,10 @@ class ParceraServer(ParceraAvatarBase):
 
         # Sync Twitch Client
         try:
-            asyncio.get_running_loop().create_task(self.sync_twitch_client())
+            asyncio.get_running_loop().create_task(self.twitch_service.sync_client())
         except RuntimeError:
-            # Skip if no loop is running (e.g. during test import or CLI initialization)
+            # Skip if no loop is running
             pass
-
-    async def sync_twitch_client(self):
-        """Synchronize twitch client state with current config."""
-        if not hasattr(self, "twitch_client") or self.twitch_client is None:
-            return
-
-        settings = self.config.settings.get("twitch", {})
-
-        # 1. Update filters
-        self.twitch_client.update_settings(
-            wake_word=settings.get("wake_word"),
-            ignored_users=settings.get("ignored_users"),
-            ng_words=settings.get("ng_words")
-        )
-
-        # 2. Control Chat Listener
-        enabled = settings.get("enabled", False)
-
-        main_loop = asyncio.get_running_loop()
-
-        if enabled:
-            async def chat_callback(user_name, text):
-                chat_logger.log_twitch(user_name, text)
-
-                async def enqueue():
-                    await self.twitch_queue.put((user_name, text))
-
-                asyncio.run_coroutine_threadsafe(enqueue(), main_loop)
-
-            if not self.twitch_client.is_chat_started:
-                await self.twitch_client.start_chat(on_message=chat_callback)
-            else:
-                self.twitch_client.on_message_callback = chat_callback
-        else:
-            if self.twitch_client.is_chat_started:
-                await self.twitch_client.stop_chat()
-
-    CLEAN_TEXT_RE = re.compile(r"[^\w\u3040-\u309F\u30A0-\u30FF\u4E00-\u9FFF]")
-
-    def _calculate_twitch_wait_time(self, text: str) -> float:
-        twitch_cfg = self.config.get("twitch", {})
-        speed = twitch_cfg.get("response_speed", "natural")
-
-        presets = {
-            "instant": [0.0, 0.0],
-            "fast":    [0.1, 0.03],
-            "natural": [0.2, 0.07],
-            "slow":    [0.5, 0.12],
-        }
-        base, spw = presets.get(speed, presets["natural"])
-
-        if speed == "instant":
-            return base
-
-        # Weight: Kanji=2, others=1. Ignore symbols.
-        clean_text = self.CLEAN_TEXT_RE.sub("", text)
-        weight = sum(2 if "\u4e00" <= c <= "\u9fff" else 1 for c in clean_text)
-
-        return base + (weight * spw)
-
-    async def _process_twitch_queue(self):
-        logger.info("Twitch queue processor started.")
-        while True:
-            try:
-                user_name, text = await self.twitch_queue.get()
-                logger.debug(f"Twitch Queue: Processing message from <{user_name}>")
-
-                # 2. Start LLM invocation IMMEDIATELY (Background thinking)
-                # Even if AI is busy with the user, we can start generating the response.
-                wait_time = self._calculate_twitch_wait_time(text)
-                logger.info(f"Twitch Queue: Starting LLM background thinking for <{user_name}> (reading wait: {wait_time:.2f}s)")
-
-                # Invoke AI. It will handle the internal busy-wait and reading delay before audio starts.
-                await self._invoke_twitch_response(user_name, text, audio_delay=wait_time)
-
-                self.twitch_queue.task_done()
-
-                # Space out consecutive queued messages
-                await asyncio.sleep(0.5)
-            except Exception as e:
-                logger.error(f"Error in Twitch queue processor: {e}", exc_info=True)
-                await asyncio.sleep(1.0)
-
-    async def _invoke_twitch_response(self, user_name, text, audio_delay: float = 0.0):
-        full_text = f"[Twitch Viewer] {user_name}: {text}"
-        logger.info(f"Invoking Twitch Response (Thinking) for <{user_name}>: {text}")
-
-        start_time = asyncio.get_event_loop().time()
-        first_chunk = True
-
-        try:
-            async for r in self.aiavatar_server.sts.invoke(STSRequest(
-                type="invoke",
-                session_id=TWITCH_SESSION_ID,
-                text=full_text
-            )):
-                # If this is the first response chunk (usually TTS starting),
-                # we wait for both 'reading delay' AND 'user busy' state before speaking.
-                if first_chunk:
-                    # 1. Wait for other sessions (like parcera-session) to finish
-                    while self.is_busy(exclude_session=TWITCH_SESSION_ID):
-                        await asyncio.sleep(0.2)
-
-                    # 2. Lock the busy flag for Twitch now that we are actually about to speak
-                    self.set_busy(TWITCH_SESSION_ID, True, timeout=20.0, source="twitch")
-
-                    # 3. Ensure emulated reading wait is satisfied
-                    if audio_delay > 0:
-                        elapsed = asyncio.get_event_loop().time() - start_time
-                        remaining = audio_delay - elapsed
-                        if remaining > 0:
-                            logger.debug(f"Twitch: LLM was fast ({elapsed:.2f}s), waiting {remaining:.2f}s more.")
-                            await asyncio.sleep(remaining)
-
-                    first_chunk = False
-
-                await self.aiavatar_server.handle_response(r)
-        except Exception as e:
-            logger.error(f"Error invoking AI from Twitch Chat: {e}")
-            self.set_busy(TWITCH_SESSION_ID, False)
 
 
 # ─── Initialization ─────────────────────────────────────────
@@ -303,33 +174,31 @@ async def lifespan(app: FastAPI):
         await parcera_server.tts_engine_manager.start()
         parcera_server.current_tts_provider = provider
 
-    # Ensure runtime configs (like Twitch) are applied once the loop is definitely running
     logger.info("LIFESPAN: Applying final runtime configuration and syncing Twitch...")
     parcera_server.apply_runtime_config()
 
-    asyncio.create_task(parcera_server.warmup())
-    asyncio.create_task(parcera_server._process_twitch_queue())
+    parcera_server._warmup_task = asyncio.create_task(parcera_server.warmup())
+    parcera_server._twitch_task = asyncio.create_task(parcera_server.twitch_service.process_queue())
 
     yield
 
     # Shutdown
+    if hasattr(parcera_server, '_warmup_task'): parcera_server._warmup_task.cancel()
+    if hasattr(parcera_server, '_twitch_task'): parcera_server._twitch_task.cancel()
     await parcera_server.cleanup()
     if parcera_server.tts_engine_manager:
         await parcera_server.tts_engine_manager.stop()
+
 
 
 # ─── FastAPI App ─────────────────────────────────────────────
 
 app = FastAPI(lifespan=lifespan)
 
-
-# Middleware to gracefully handle WebSocketDisconnect that aiavatar library doesn't catch.
-# Without this, normal client disconnections (page reloads, reconnects) log as ERROR in uvicorn.
 from starlette.websockets import WebSocketDisconnect
 
 class WebSocketDisconnectMiddleware:
     """ASGI middleware that catches WebSocketDisconnect to prevent noisy ERROR logs."""
-
     def __init__(self, app):
         self.app = app
 
@@ -371,74 +240,11 @@ async def health_check():
     return {"status": "ok", "tts_engine": tts_ok}
 
 
-# --- Model Management Endpoints ---
-
-
-@app.get("/models/check")
-async def check_model(name: str):
-    """Check if a Whisper model is already cached."""
-    cached = check_model_cached(name)
-    return {"cached": cached, "model": name}
-
-
-@app.get("/models/download")
-async def download_model_sse(name: str):
-    """Download a Whisper model with SSE progress streaming."""
-
-    async def event_stream() -> AsyncIterator[str]:
-        loop = asyncio.get_running_loop()
-        done_event = asyncio.Event()
-        result = {"error": None}
-
-        def _download():
-            try:
-                download_model_with_progress(name)
-            except Exception as e:
-                result["error"] = str(e)
-                logger.error(f"Download thread error: {e}")
-            finally:
-                loop.call_soon_threadsafe(done_event.set)
-
-        # Start download in a thread (it's blocking I/O)
-        thread = threading.Thread(target=_download, daemon=True)
-        thread.start()
-
-        last_pct = -1
-        try:
-            while not done_event.is_set():
-                progress = LoggingTqdm.get_progress()
-                if progress and progress["progress"] != last_pct:
-                    last_pct = progress["progress"]
-                    yield f"data: {json.dumps(progress)}\n\n"
-                await asyncio.sleep(0.5)
-
-            # Final event
-            if result["error"]:
-                yield f"data: {json.dumps({'progress': -1, 'status': 'error', 'error': result['error']})}\n\n"
-            else:
-                yield f"data: {json.dumps({'progress': 100, 'status': 'complete'})}\n\n"
-
-        finally:
-            LoggingTqdm.reset_progress()
-
-    return StreamingResponse(event_stream(), media_type="text/event-stream")
-
-
-@app.post("/models/reload")
-async def reload_model():
-    """Trigger a reload of the STT component."""
-    try:
-        parcera_server.reload_stt()
-        return {"success": True}
-    except Exception as e:
-        logger.error(f"Failed to reload STT: {e}")
-        return {"success": False, "error": str(e)}
-
-
 # Register routers
 app.include_router(create_config_router(_get_server))
 app.include_router(create_tts_router(_get_server))
 app.include_router(create_twitch_router(_get_server))
+app.include_router(create_model_router(_get_server))
 app.include_router(parcera_server.aiavatar_server.get_websocket_router())
 
 
@@ -446,5 +252,5 @@ if __name__ == "__main__":
     import uvicorn
     config_path = os.environ.get("PARCERA_CONFIG_PATH", "configs/settings.default.yaml")
     settings = load_config_file(config_path)
-    port = settings.get("electron", {}).get("port", 8676)
-    uvicorn.run(app, host="127.0.0.1", port=port)
+    port = settings.get("electron", {}).get("port", DEFAULT_UVICORN_PORT)
+    uvicorn.run(app, host=DEFAULT_UVICORN_HOST, port=port)
